requirements.txt
pandas
python-pptx
Pillow
